Readme: update speed comparison
First of all, thanks for bringing a new life to PhlexUI (I've been using a custom fork for a while; now there is a hope that I can switch to the new upstream some day 🙂).

I've been checking the repo recently and found that the speed comparison is out of tune. Phlex is fast but the difference is not that big (compared to the primary competitor—View Component).

I've updated the numbers to reflect the latest results from the Konnor's repo as well as [my modification of it](https://github.com/palkan/view-layer-benchmarks/tree/table-component) (I tried to test against a real-world example—and that's a PhlexUI Table component, btw).

Signed-off-by: Vladimir Dementyev <dementiev.vm@gmail.com>