Add generator for all components (#289)
* Add generator for all components

Add a new generator that allows users to generate all components at once
using the command `bin/rails g ruby_ui:component:all`.

This makes it easier for users who want to install the complete set of
components without having to run individual commands for each component.

In my notebook, it's tanking 1m 3s to generate all components. We can improve
this but i tried to keep the same approach as the component generator to
make it easier to maintain.

* fix linter

* add force option to all generators

* Change all generator to use force option

* Add 3.4 to CI and remove head

I was getting a segmentation fault when running the tests on head